requirements.txt
alembic
aiosqlite==0.21.0
asyncpg==0.29.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cache2
greenlet==3.1.1
h11==0.14.0
httpx
idna==3.10
kombu==5.5.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.6
pydantic==2.10.6
pydantic_core==2.27.2
pydantic-settings
python-dateutil==2.9.0.post0
pytest
pytest-asyncio
redis==5.2.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy[asyncio]==2.0.39
starlette==0.46.1
typing_extensions
tzdata==2025.1
unicorn==2.1.3
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
